parsing-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pytesseract==0.3.13
easyocr==1.7.2
img2table==1.3.0
camelot-py[cv]==0.11.0
Pillow==11.0.0
numpy==2.1.0
opencv-python-headless==4.10.0.84
tika==2.6.0
openpyxl==3.1.5
pydantic==2.10.0
